requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
assertpy==1.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
